selesai dalam migrasi model factories dan seeder dengan patokan ERD part1

+++ database/migrations/2026_04_01_000005_create_shift_table.php
@@ -9,7 +9,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('shift', function (Blueprint $table) {
-            $table->integer('id_shift')->autoIncrement();
+            $table->bigInteger('id_shift')->autoIncrement();
             $table->time('jam_masuk');
             $table->time('jam_keluar');
             $table->string('tipe_shift', 50);
