+++ database/migrations/2026_04_01_000010_create_tipe_kehadiran_table.php
@@ -9,7 +9,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('tipe_kehadiran', function (Blueprint $table) {
-            $table->integer('id_tipe_kehadiran')->autoIncrement();
+            $table->bigInteger('id_tipe_kehadiran')->autoIncrement();
             $table->enum('status_kehadiran', ['hadir', 'sakit', 'izin', 'mankir', 'cuti', 'terlambat']);
             $table->timestamps();
         });
